Route Planner with Bing MAP
I have created this app based on the requirements for my wife to allow lookup of distance to from client locations as she is working as a Care Assistance which requires her to visit to multiple elderly per shift (10-12)

The idea is that you use a simple entry form to type up client details for future use and store it in a database which then can be displayed and selected this functionality only works with the db i have on her laptop can be fixed by pointing the connection to another database and creating a few entries there.

Also there is an error Object reference is not set to an instance in the map tab this can be ignored and the displayed result seems to be correct - not surpressed it yet.